adding readme for retail success, supporting repositories with three part names, adding interfaces, ignoring status endpoint, supporting cancellation token, supporting enumerable query string parameters (along with sdk core change), using decimal instead of double, removing non-async wrapper methods, adding powershell script to set up for image generation

+++ src/vanilla/Model/PrimaryTypeCs.cs
@@ -54,7 +54,7 @@ namespace AutoRest.CSharp.Model
                         return "System.DateTime";
 
                     case KnownPrimaryType.Double:
-                        return "double";
+                        return "decimal";
 
                     case KnownPrimaryType.Decimal:
                         return "decimal";
